+++ src/strategy/fixedweight.rs
@@ -25,7 +25,7 @@ impl Strategy for FixedWeightStrategy {
         self.perf.update(&state)
     }
 
-    fn init(&mut self, initital_cash: &f64) {
+    fn init(&mut self, initital_cash: &u64) {
         self.portfolio.deposit_cash(initital_cash);
     }
