+++ src/strategy/mod.rs
@@ -16,6 +16,6 @@ pub mod staticweight;
 pub trait Strategy {
     fn run(&mut self) -> f64;
     fn set_date(&mut self, date: &i64);
-    fn init(&mut self, initial_cash: &f64);
+    fn init(&mut self, initial_cash: &u64);
     fn get_perf(&self) -> PerfStruct;
 }
